chore: Improve a11y of loading state of button (#23812)
Fixes #23811
