+++ app/client/packages/design-system/widgets/package.json
@@ -15,6 +15,7 @@
     "@design-system/headless": "workspace:^",
     "@design-system/theming": "workspace:^",
     "@react-aria/utils": "^3.16.0",
+    "@react-aria/visually-hidden": "^3.8.0",
     "colorjs.io": "^0.4.3"
   },
   "devDependencies": {
