+++ app/client/packages/design-system/widgets/src/components/Button/Button.stories.mdx
@@ -76,6 +76,7 @@ There are 3 variants of the button component.
     args={{
       isLoading: true,
       children: "Loading",
+      onPress: () => alert("on pressed"),
     }}
   >
     {Template.bind({})}
