+++ app/client/packages/design-system/widgets/src/components/Button/Button.tsx
@@ -1,14 +1,15 @@
 import React, { forwardRef } from "react";
-import { Icon as HeadlessIcon } from "@design-system/headless";
+import { useVisuallyHidden } from "@react-aria/visually-hidden";
+
+import { Text } from "../Text";
 import type {
   ButtonProps as HeadlessButtonProps,
   ButtonRef as HeadlessButtonRef,
 } from "@design-system/headless";
-
-import { Text } from "../Text";
 import { Spinner } from "../Spinner";
 import { StyledButton } from "./index.styled";
 import type { fontFamilyTypes } from "../../utils/typography";
+import { Icon as HeadlessIcon } from "@design-system/headless";
 
 export type ButtonVariants = "primary" | "secondary" | "tertiary";
 
@@ -17,16 +18,18 @@ export interface ButtonProps extends Omit<HeadlessButtonProps, "className"> {
    *  @default primary
    */
   variant?: ButtonVariants;
-  isLoading?: boolean;
   fontFamily?: fontFamilyTypes;
   isFitContainer?: boolean;
   isFocused?: boolean;
+  isLoading?: boolean;
   icon?: React.ReactNode;
   iconPosition?: "start" | "end";
+  visuallyDisabled?: boolean;
 }
 
 export const Button = forwardRef(
   (props: ButtonProps, ref: HeadlessButtonRef) => {
+    props = useVisuallyDisabled(props);
     const {
       children,
       fontFamily,
@@ -34,18 +37,23 @@ export const Button = forwardRef(
       iconPosition = "start",
       isFitContainer = false,
       isLoading,
-      // eslint-disable-next-line -- TODO add onKeyUp when the bug is fixedhttps://github.com/adobe/react-spectrum/issues/4350
+      // eslint-disable-next-line -- TODO add onKeyUp when the bug is fixed https://github.com/adobe/react-spectrum/issues/4350
       onKeyUp,
       variant = "primary",
+      visuallyDisabled,
       ...rest
     } = props;
+    const { visuallyHiddenProps } = useVisuallyHidden();
 
     const renderChildren = () => {
       if (isLoading) {
         return (
-          <HeadlessIcon>
-            <Spinner />
-          </HeadlessIcon>
+          <>
+            <HeadlessIcon>
+              <Spinner />
+            </HeadlessIcon>
+            <span {...visuallyHiddenProps}>Loading...</span>
+          </>
         );
       }
 
@@ -61,6 +69,8 @@ export const Button = forwardRef(
 
     return (
       <StyledButton
+        aria-busy={isLoading ? true : undefined}
+        aria-disabled={visuallyDisabled || isLoading ? true : undefined}
         data-button=""
         data-fit-container={isFitContainer ? "" : undefined}
         data-icon-position={iconPosition === "start" ? undefined : "end"}
@@ -74,3 +84,31 @@ export const Button = forwardRef(
     );
   },
 );
+
+/**
+ * This hook is used to disable all click/press events on a button
+ * when the button is visually disabled
+ *
+ * @param props
+ * @returns
+ */
+const useVisuallyDisabled = (props: ButtonProps) => {
+  let computedProps = props;
+
+  if (props.visuallyDisabled || props.isLoading) {
+    computedProps = {
+      ...props,
+      isDisabled: false,
+      // disabling click/press events
+      onPress: undefined,
+      onPressStart: undefined,
+      onPressEnd: undefined,
+      onPressChange: undefined,
+      onPressUp: undefined,
+      onKeyDown: undefined,
+      onKeyUp: undefined,
+    };
+  }
+
+  return computedProps;
+};
