+++ app/client/packages/design-system/widgets/src/components/Button/index.styled.tsx
@@ -20,20 +20,16 @@ export const StyledButton = styled(HeadlessButton)<ButtonProps>`
     height: 100%;
   }
 
-  &[data-loading="true"] {
-    pointer-events: none;
-  }
-
   &[data-variant="primary"] {
     background-color: var(--color-bg-accent);
     color: var(--color-fg-on-accent);
     border-color: transparent;
 
-    &[data-hovered] {
+    &[data-hovered]:not(:is([data-loading], [data-disabled], [aria-disabled])) {
       background-color: var(--color-bg-accent-hover);
     }
 
-    &[data-active] {
+    &[data-active]:not(:is([data-loading], [data-disabled], [aria-disabled])) {
       background-color: var(--color-bg-accent-active);
     }
   }
@@ -44,11 +40,11 @@ export const StyledButton = styled(HeadlessButton)<ButtonProps>`
     border-color: var(--color-bd-accent);
     border-width: var(--border-width-1);
 
-    &[data-hovered] {
+    &[data-hovered]:not(:is([data-loading], [data-disabled], [aria-disabled])) {
       background-color: var(--color-bg-accent-subtle-hover);
     }
 
-    &[data-active] {
+    &[data-active]:not(:is([data-loading], [data-disabled], [aria-disabled])) {
       background-color: var(--color-bg-accent-subtle-active);
     }
   }
@@ -59,23 +55,28 @@ export const StyledButton = styled(HeadlessButton)<ButtonProps>`
     border-color: transparent;
     border-width: 0;
 
-    &[data-hovered] {
+    &[data-hovered]:not(:is([data-loading], [data-disabled], [aria-disabled])) {
       background: var(--color-bg-accent-subtle-hover);
     }
 
-    &[data-active] {
+    &[data-active]:not(:is([data-loading], [data-disabled], [aria-disabled])) {
       background: var(--color-bg-accent-subtle-active);
     }
   }
 
-  &[data-focused]:not([data-loading]) {
+  &[data-focused] {
     box-shadow: 0 0 0 2px var(--color-bg), 0 0 0 4px var(--color-bd-focus);
   }
 
   &[data-disabled],
-  &[aria-disabled] {
+  // TODO(Pawan): Check if we need to make the UI of loading state same as disabled?
+  &[aria-disabled]:not([data-loading]) {
     cursor: default;
     opacity: var(--opacity-disabled);
+  }
+
+  &[data-loading] {
+    cursor: default;
   }
 
   & [data-icon] {
